Ignore team_rename RTM event for multiple teams
These events get fired for shared teams too and it's impossible to
identify which team renamed.
Only way to refresh a team name is using the events api or restart
bridge.

From Slack docs:

Clients can use this to update the display of the workspace name as
soon as it changes. If they don't the client will receive the new name
the next time it calls rtm.start.

Could be altered to update all teams using team.info methods.